@@ -20,7 +20,8 @@
 class Object3D
 {
 public:
-	Object3D() = default;
+	Object3D(std::string_view name = "New Object")
+		: m_Name(name) {}
 	~Object3D() = default;
 
 	void Init(const std::vector<float> &vertices, const std::vector<uint32_t> &indices, BufferLayout layout);
@@ -28,6 +29,9 @@ public:
 	// const std::shared_ptr<VertexArray> &GetVertexArray() const { return m_VertexArray; }
 
 	UUID GetID() const { return m_ID; }
+
+	const std::string& GetName() const { return m_Name; }
+	void SetName(const std::string &name) { m_Name = name; }
 
 	glm::vec3 GetPosition() const { return m_Position; }
 	void SetPosition(const glm::vec3 &position)
@@ -50,32 +54,56 @@ public:
 		m_TransformDirty = true;
 	}
 
+	glm::mat4 GetWorldTransform() const
+	{
+		// TODO: this is very temporary
+		glm::mat4 transform = glm::mat4(1.0f);
+
+		// Apply translation
+		transform = glm::translate(transform, m_Position);
+
+		// Apply rotation (in order: Z, Y, X)
+		// transform = glm::rotate(transform, m_Rotation.z, glm::vec3(0.0f, 0.0f, 1.0f));
+		// transform = glm::rotate(transform, m_Rotation.y, glm::vec3(0.0f, 1.0f, 0.0f));
+		// transform = glm::rotate(transform, m_Rotation.x, glm::vec3(1.0f, 0.0f, 0.0f));
+
+		// If m_Rotation is in degrees, convert to radians first
+		transform = glm::rotate(transform, glm::radians(m_Rotation.z), glm::vec3(0.0f, 0.0f, 1.0f));
+		transform = glm::rotate(transform, glm::radians(m_Rotation.y), glm::vec3(0.0f, 1.0f, 0.0f));
+		transform = glm::rotate(transform, glm::radians(m_Rotation.x), glm::vec3(1.0f, 0.0f, 0.0f));
+
+		// Apply scale
+		transform = glm::scale(transform, m_Scale);
+
+		return transform;
+	}
+
 	std::shared_ptr<Mesh> GetMesh() const { return m_Mesh; }
 	void SetMesh(const std::shared_ptr<Mesh> &mesh)
 	{
 		m_Mesh = mesh;
 	}
 
-	static std::shared_ptr<Object3D> Create(const std::string &name = "DefaultObject3D")
+	static std::shared_ptr<Object3D> Create(std::string_view name = "New Object")
 	{
-		return std::make_shared<Object3D>();
+		return std::make_shared<Object3D>(name);
 	}
 
-	static std::shared_ptr<Object3D> CreateCube(const std::string &name = "DefaultObject3D")
+	static std::shared_ptr<Object3D> CreateCube(std::string_view name = "New Object")
 	{
-		auto object = std::make_shared<Object3D>();
+		auto object = std::make_shared<Object3D>(name);
 		object->SetMesh(Mesh::CreateDefaultCube());
 		return object;
 	}
-	static std::shared_ptr<Object3D> CreateSphere(const std::string &name = "DefaultObject3D")
+	static std::shared_ptr<Object3D> CreateSphere(std::string_view name = "New Object")
 	{
-		auto object = std::make_shared<Object3D>();
+		auto object = std::make_shared<Object3D>(name);
 		object->SetMesh(Mesh::CreateDefaultSphere());
 		return object;
 	}
-	static std::shared_ptr<Object3D> CreateTriangle(const std::string &name = "DefaultObject3D")
+	static std::shared_ptr<Object3D> CreateTriangle(std::string_view name = "New Object")
 	{
-		auto object = std::make_shared<Object3D>();
+		auto object = std::make_shared<Object3D>(name);
 		object->SetMesh(Mesh::CreateDefaultTriangle());
 		return object;
 	}
@@ -96,6 +124,7 @@ public:
 private:
 	//? Identity
 	UUID m_ID;
+	std::string m_Name = "DefaultObject3D";
 
 	//? Transform
 	glm::vec3 m_Position = glm::vec3(0.0f, 0.0f, 0.0f);
@@ -117,7 +146,6 @@ private:
 	std::weak_ptr<Object3D> m_Parent;
 	std::vector<std::shared_ptr<Object3D>> m_Children;
 	// std::weak_ptr<Scene> m_Scene;
-
 };
 
 // class Camera : public Object3D
