@@ -22,25 +22,42 @@ class Scene
 {
 public:
 	Scene();
-	virtual ~Scene();
-
-	// virtual void OnAttach() {}
-	// virtual void OnDetach() {}
-	// virtual void OnUpdate(float deltaTime) {}
-	// virtual void OnImGuiRender() {}
-	// virtual void OnEvent(Event& event) {}
+	~Scene();
 
 	void Update(float deltaTime);
 
-	void OnImGuiRender();
+	void RenderUI();
 
+	void PrepMeshBuffers();
+
+	void RenderViewport();
 	void Render();
+
+	uint32_t GetObjectCount() const { return static_cast<uint32_t>(m_RootObjects.size()); }
 
 	void SetMesh(const std::shared_ptr<Mesh> &mesh) { m_CubeMesh = mesh; }
 	void SetEditorCamera(const std::shared_ptr<EditorCamera> &editorCamera) { m_EditorCamera = editorCamera; }
 
 	// inline const std::string& GetName() const { return m_DebugName; }
 private:
+	struct MeshInfo
+	{
+		glm::mat4 invModelMatrix;
+		uint32_t positionOffset; // Offset into the positions array
+		uint32_t normalOffset;	 // Offset into the normals array
+		uint32_t indexOffset;	 // Offset into the indices array
+		uint32_t numVertices;	 // Number of vertices in this mesh
+		uint32_t numIndices;	 // Number of indices in this mesh
+		uint32_t numFaces;		 // Number of triangles (numIndices / 3)
+		uint32_t bvhNodeOffset;	 // Offset into the BVH nodes array for this mesh
+		uint32_t aabbOffset;	 // Offset into the AABB array for this mesh
+	};
+
+	static_assert(sizeof(MeshInfo) == 96, "MeshInfo size mismatch!");
+
+private:
+	std::vector<MeshInfo> m_MeshInfos;
+
 	std::string m_Name = "Default Scene";
 	std::shared_ptr<MaterialLibrary> m_MaterialLibrary = std::make_shared<MaterialLibrary>();
 	std::shared_ptr<ShaderLibrary> m_ShaderLibrary = std::make_shared<ShaderLibrary>();
@@ -49,9 +66,9 @@ private:
 
 	std::shared_ptr<EditorCamera> m_EditorCamera;
 
+
+	std::shared_ptr<Object3D> m_SelectedObject = nullptr;
 	std::vector<std::shared_ptr<Object3D>> m_RootObjects;
-	// std::unordered_map<uint32_t, std::weak_ptr<Object3D>> m_ObjectsById;
-	// std::unordered_map<std::string, std::weak_ptr<Object3D>> m_ObjectsByName;
 
 	ShaderStorageBuffer m_AABBBuffer;
 	uint32_t m_AABBBufferCount = 0;
@@ -65,4 +82,7 @@ private:
 	std::shared_ptr<Shader> m_Shader;
 	std::shared_ptr<ComputeShader> m_ComputeShader = std::make_shared<ComputeShader>();
 	std::shared_ptr<ComputeShader> m_ComputeCalcMinMax = std::make_shared<ComputeShader>();
+
+
+	bool m_VisualizeBLAS = true;
 };
